+++ bluedragon-distribution-worker/src/main/java/com/jd/bluedragon/distribution/consumer/jy/vehicle/JyUnloadCarAggsConsumer.java
@@ -57,7 +57,7 @@ public class JyUnloadCarAggsConsumer extends MessageBaseConsumer {
         //过滤旧版本数据
         String versionMutex = String.format(CacheKeyConstants.JY_UNLOAD_AGG_KEY, entity.getBizId()+entity.getProductType());
         if (redisClientOfJy.exists(versionMutex)) {
-            Integer version = Integer.valueOf(redisClientOfJy.get(versionMutex));
+            Long version = Long.valueOf(redisClientOfJy.get(versionMutex));
             if (!NumberHelper.gt(entity.getVersion(), version)) {
                 logger.warn("JyUnloadCarAggsConsumer receive old version data. curVersion: {}, 内容为【{}】", version, message.getText());
                 return;
